+++ lib/src/tools/em_extension.dart
@@ -1,4 +1,3 @@
-import '../internal/inner_headers.dart';
 import 'dart:convert' as convert;
 
 Type typeOf<T>() => T;
@@ -20,112 +19,11 @@ extension MapExtension on Map {
     }
   }
 
-  int? intValue(String key) {
-    if (this.containsKey(key)) {
-      return this[key];
-    } else {
-      return null;
+  void add(String key, dynamic value) {
+    if (value == null) {
+      return;
     }
-  }
-
-  String? stringValue(String key) {
-    if (this.containsKey(key)) {
-      return this[key];
-    } else {
-      return null;
-    }
-  }
-
-  List<T>? listValue<T>(String key) {
-    if (this.containsKey(key)) {
-      List obj = this[key];
-      if (typeOf<T>().toString() == "String") {
-        List<String> strList = [];
-        for (var item in obj) {
-          strList.add(item);
-        }
-        return strList as List<T>;
-      } else if (typeOf<T>().toString() == "EMGroupSharedFile") {
-        List<EMGroupSharedFile> fileList = [];
-        for (var item in obj) {
-          var file = EMGroupSharedFile.fromJson(item);
-          fileList.add(file);
-        }
-        return fileList as List<T>;
-      } else if (typeOf<T>().toString() == "EMPresence") {
-        List<EMPresence> presenceList = [];
-        for (var item in obj) {
-          var presence = EMPresence.fromJson(item);
-          presenceList.add(presence);
-        }
-        return presenceList as List<T>;
-      }
-    }
-    return null;
-  }
-
-  ///
-  /// 如果给的value是null则不设置到map中。
-  void setValueWithOutNull<T>(String key, T? value,
-      {Object Function(T object)? callback, T? defaultValue}) {
-    if (value != null) {
-      if (callback != null) {
-        Object v = callback(value);
-        this[key] = v;
-      } else {
-        this[key] = value;
-      }
-    } else {
-      if (defaultValue != null) {
-        this[key] = defaultValue;
-      }
-    }
-  }
-
-  int? getIntValue(String key, {int? defaultValue}) {
-    int? ret = defaultValue;
-    if (this.containsKey(key)) {
-      dynamic value = this[key];
-      if (value is int) {
-        ret = value;
-      }
-    }
-    return ret;
-  }
-
-  bool? getBoolValue(String key, {bool? defaultValue}) {
-    bool? ret = defaultValue;
-    if (this.containsKey(key)) {
-      dynamic value = this[key];
-      if (value is bool) {
-        ret = value;
-      }
-    }
-    return ret;
-  }
-
-  String? getStringValue(String key, {String? defaultValue}) {
-    String? ret = defaultValue;
-    if (this.containsKey(key)) {
-      dynamic value = this[key];
-      if (value is String) {
-        ret = value;
-      }
-    }
-    return ret;
-  }
-
-  double? getDoubleValue(String key, {double? defaultValue}) {
-    double? ret = defaultValue;
-    if (this.containsKey(key)) {
-      dynamic value = this[key];
-      if (value is double) {
-        ret = value;
-      } else if (value is int) {
-        ret = value.toDouble();
-      }
-    }
-    return ret;
+    this[key] = value;
   }
 
   Map? getMapValue(String key, {Map? defaultValue}) {
@@ -154,7 +52,7 @@ extension MapExtension on Map {
     return ret;
   }
 
-  List<T>? getList<T>(String key, {valueCallback: MapResultCallback}) {
+  List<T>? getList<T>(String key, {valueCallback = MapResultCallback}) {
     List<T>? ret;
     if (this.containsKey(key)) {
       List list = this[key];
@@ -170,7 +68,7 @@ extension MapExtension on Map {
     return ret;
   }
 
-  T? getValueWithKey<T>(
+  T? getValue<T>(
     String key, {
     required MapResultCallback callback,
   }) {